code structure first draft
I read the directions provided and without coding any function in laid out the approximate structure of the program. Next task is to go in and add functionality to each method in turn. I don't think I need available funds when I instantiate the categories, but I am leaving it there for now.